requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
motor>=3.6.0
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
clip @ git+https://github.com/openai/CLIP.git@ded190a052fdf4585bd685cee5bc96e0310d2c93
colorama==0.4.6
cryptography==46.0.5
distro==1.9.0
dnspython==2.8.0
filelock==3.24.3
fsspec==2026.2.0
ftfy==6.3.1
google-ai-generativelanguage==0.6.15
google-api-core==2.30.0
google-api-python-client==2.190.0
google-auth==2.49.0.dev0
google-auth-httplib2==0.3.0
google-genai==1.65.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
grpcio==1.78.1
grpcio-status==1.71.2
h11==0.16.0
h2==4.3.0
hf-xet==1.3.1
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
huggingface_hub==1.4.1
hyperframe==6.1.0
idna==3.11
ImageHash==4.3.2
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.2
packaging==26.0
pillow==12.1.1
portalocker==3.2.0
proto-plus==1.27.1
protobuf==5.29.6
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pymongo==4.16.0
pyparsing==3.3.2
python-dotenv==1.2.1
PyWavelets==1.9.0
pywin32==311
PyYAML==6.0.3
qdrant-client==1.17.0
regex==2026.2.19
requests==2.32.5
rich==14.3.3
safetensors==0.7.0
scipy==1.17.1
setuptools==82.0.0
shellingham==1.5.4
sniffio==1.3.1
sympy==1.14.0
tenacity==9.1.4
tokenizers==0.22.2
torch==2.10.0
torchvision==0.25.0
tqdm==4.67.3
transformers==5.2.0
typer==0.24.1
typer-slim==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
wcwidth==0.6.0
websockets==16.0
